remove arquivo utils para corrigir problema de dependencia circular. As funções foram movidas para o UserService na pasta services
