+++ back-end/src/controllers/RegisterController.js
@@ -1,10 +1,9 @@
-const { create } = require('../services/UserService');
-const { generateJWTToken } = require('../utils');
+const { generateJWTToken, create } = require('../services/UserService');
 
 module.exports = async (req, res) => {
   const newUserInfo = req.body;
   const user = await create(newUserInfo);
-  if (user.error) return res.status(user.error.status).json(user.error.message);
+  if (user.error) return res.status(user.error.status).json({ message: user.error.message });
   const token = generateJWTToken(newUserInfo);
   return res.status(201).json({ ...user, token });
 };
