+++ back-end/src/services/UserService.js
@@ -1,11 +1,21 @@
 const md5 = require('md5');
+const jwt = require('jsonwebtoken');
+const fs = require('fs');
+
 const { User } = require('../database/models');
-const { validateNewUserInfo } = require('../utils');
+
+const generateJWTToken = (user) => {
+  const jwtConfig = { expiresIn: '7d', algorithm: 'HS256' };
+  const secret = fs.readFileSync('./jwt.evaluation.key', 'utf-8');
+  console.log('secret: ', secret);
+  const { name, email, role } = user;
+  const token = jwt.sign({ data: { name, email, role } }, secret, jwtConfig);
+  return token;
+};
 
 const findByName = async (name) => {
   const user = await User.findOne({ where: { name } });
-  const { password, ...data } = user;
-  return data;
+  return user;
 };
 
 const findByEmail = async (email) => {
@@ -13,12 +23,23 @@ const findByEmail = async (email) => {
   return user;
 };
 
-const create = async (newUserInfo) => {
- const isValid = validateNewUserInfo(newUserInfo);
- if (!isValid) return { error: { status: 409, message: 'Name or email already registered' } };
- const { name, email, password } = newUserInfo;
- await User.create({ name, email, password: md5(password), role: 'customer' });
- return { name, email, role: 'customer' };
+const validateNewUserInfo = async (newUserInfo) => {
+  const { name, email } = newUserInfo;
+  const nameExists = await findByName(name);
+  if (nameExists) return false;
+  const emailExists = await findByEmail(email);
+  if (emailExists) return false;
+  return true;
 };
 
-module.exports = { findByEmail, findByName, create };
+const create = async (newUserInfo) => {
+  const isValid = await validateNewUserInfo(newUserInfo);
+  if (!isValid) {
+    return { error: { status: 409, message: 'Name or email already registered' } };
+  }
+  const { name, email, password } = newUserInfo;
+  await User.create({ name, email, password: md5(password), role: 'customer' });
+  return { name, email, role: 'customer' };
+};
+
+module.exports = { generateJWTToken, findByEmail, findByName, create };
